improve lesson page //admin edit lesson

+++ pages/admin/Delete.php
@@ -22,23 +22,45 @@ elseif (isset($_GET['lessId'])) {
     $sh_id = $_POST['sh_delete'];
 }
 
-
+///////// set ตัวแปลส่วนของหน้า คำศัพท์
 if (isset($_GET['vocabId'])) {
     $query = "DELETE FROM $hsk WHERE $HSK_id = '$vocabId' ";
     $result = mysqli_query($conn, $query);
     $_SESSION['delete'] = "ลบคำศัพท์สำเร็จ";
     header('location:./editVocab.php?set=' . $_GET['set'] . '&&part=' . $_GET['part']);
-} elseif (isset($_GET['lessId'])) {
-    $query = "DELETE FROM $hsk WHERE $HSK_id = '$lessId' ";
-    $result = mysqli_query($conn, $query);
-    $_SESSION['delete'] = "ลบบทเรียนสำเร็จ";
-    header('location:./editLess.php?set=' . $_GET['set'] . '&&part=' . $_GET['part']);
-} elseif ($tid) {
+}
+///////// set ตัวแปลส่วนของหน้า บทเรียน 
+elseif (isset($_GET['lessId'])) {
+    $name_img = $_GET['name_img'];
+    $name_sound = $_GET['name_sound'];
+
+    $locate_img = '../../img/บทเรียนHSK' . $_GET['set'] . '_ชุดที่' . $_GET['part'] . '/';
+    $locate_sound = '../../sound/HSK' . $_GET['set'] . '_' . $_GET['part'] . '/';
+
+    $file_img = $locate_img . $name_img;
+    $file_sound = $locate_sound . $name_sound;
+
+    if (unlink($file_img) && unlink($file_sound)) {
+        $query = "DELETE FROM $hsk WHERE $HSK_id = '$lessId' ";
+        $result = mysqli_query($conn, $query);
+        $_SESSION['delete'] = "ลบบทเรียนสำเร็จ";
+        header('location:./editLess.php?set=' . $_GET['set'] . '&&part=' . $_GET['part']);
+    } else {
+        $_SESSION['delete'] = "ลบบทเรียนไม่สำเร็จ !";
+        header('location:./editLess.php?set=' . $_GET['set'] . '&&part=' . $_GET['part']);
+    }
+
+   
+}
+///////// set ตัวแปลส่วนของหน้า techer
+elseif ($tid) {
     $query = "DELETE FROM Teacher WHERE TID = '$tid' ";
     $result = mysqli_query($conn, $query);
     $_SESSION['delete'] = "ลบข้อมูลสำเร็จ";
     header('location:./teacher.php');
-} elseif ($sh_id) {
+}
+///////// set ตัวแปลส่วนของหน้าโรงเรียน
+elseif ($sh_id) {
     $query = "DELETE FROM School WHERE SchoolID = '$sh_id' ";
     $result = mysqli_query($conn, $query);
     $_SESSION['delete'] = "ลบข้อมูลสำเร็จ";
